requirements.txt
dnspython>=2.4.0
fastapi>=0.104.0
uvicorn>=0.23.2
typer>=0.9.0
rich>=13.6.0
python-whois>=0.8.0
graphviz>=0.20.1
